chore: Milestone 1 cleanup — package CLI, console entry, shim deprecations, CI polish, docs, and migrating guide
