+++ cli.py
@@ -1,637 +1,21 @@
 #!/usr/bin/env python3
-import argparse
-import json
-import os
-import shlex
-import sys
-from typing import Any
-
-# Re-export CommandExecutor so tests can patch cli.CommandExecutor
-from command_executor import CommandExecutor  # noqa: F401
-
-# Re-export run_app and ConfigCLI at module level so tests can patch cli.run_app and cli.ConfigCLI
-try:
-    from main import run_app as run_app  # type: ignore # noqa: F401
-except Exception:
-
-    def run_app() -> None:  # type: ignore
-        pass
-
-
-try:
-    from config_cli import ConfigCLI as ConfigCLI  # type: ignore # noqa: F401
-except Exception:
-
-    class ConfigCLI:  # type: ignore
-        @staticmethod
-        def interactive_mode() -> int:
-            return 0
-
-        @staticmethod
-        def list_config() -> int:
-            return 0
-
-        @staticmethod
-        def set_listen_for(key: str, value: str) -> int:
-            return 0
-
-        @staticmethod
-        def set_mode(mode: str, value: str) -> int:
-            return 0
-
-        @staticmethod
-        def run_wizard() -> int:
-            return 0
-
-        @staticmethod
-        def set_model_action(model: str, action: str) -> int:
-            return 0
-
-
-class HelpfulArgumentParser(argparse.ArgumentParser):
-    def error(self, message):
-        self.print_usage(sys.stderr)
-        self.exit(2, f"{self.prog}: error: {message}\n")
-
-
-def _get_model_actions_from_config(cfg: Any) -> dict[str, Any]:
-    try:
-        if hasattr(cfg, "__dict__"):
-            v = cfg.__dict__.get("model_actions")
-            if isinstance(v, dict):
-                return v
-    except Exception:
-        pass
-    try:
-        v = getattr(cfg, "model_actions", {})
-        if isinstance(v, dict):
-            return v
-    except Exception:
-        pass
-    return {}
-
-
-def _print_actions_text(actions: dict[str, Any]) -> None:
-    if not actions:
-        print("No commands configured.")
-        return
-    print("Available commands:")
-    for name in sorted(actions.keys()):
-        print(f"- {name}")
-
-
-def _print_actions_json(actions: dict[str, Any]) -> None:
-    arr: list[dict[str, str | None]] = []
-    for k, v in actions.items():
-        vtype = None
-        if isinstance(v, dict) and v:
-            try:
-                vtype = next(iter(v.keys()))
-            except Exception:
-                vtype = None
-        arr.append({"name": k, "type": vtype})
-    print(json.dumps(arr))
-
-
-def build_parser() -> argparse.ArgumentParser:
-    parser = HelpfulArgumentParser(prog="chatty-commander", description="ChattyCommander CLI")
-
-    if "DISPLAY" not in os.environ:
-        os.environ["DISPLAY"] = ":0"
-
-    subparsers = parser.add_subparsers(dest="command", required=False)
-
-    # Expose parser instance on itself so handlers can call parser.error(...) and trigger SystemExit
-    # This keeps behavior aligned with argparse UX and lets tests assert on stderr.
-    parser._self = parser  # type: ignore[attr-defined]
-
-    # run
-    run_parser = subparsers.add_parser(
-        "run", help="Run the main application.", description="Launch ChattyCommander core runtime."
-    )
-    run_parser.add_argument("--display", help="Override DISPLAY for GUI features.")
-
-    def run_func() -> None:
-        # Prefer patched cli.run_app if tests patched it
-        try:
-            from cli import run_app as _patched  # self-import picks patched symbol
-
-            if callable(_patched):
-                _patched()
-                return
-        except Exception:
-            pass
-        try:
-            from main import run_app as _run_app  # type: ignore
-
-            if callable(_run_app):
-                _run_app()
-        except Exception:
-            return
-
-    run_parser.set_defaults(func=run_func)
-
-    # gui
-    gui_parser = subparsers.add_parser(
-        "gui", help="Launch GUI mode.", description="Open the graphical user interface."
-    )
-
-    def gui_func(args: argparse.Namespace) -> None:
-        try:
-            from gui import run_gui  # noqa
-
-            run_gui()
-        except Exception:
-            # Allow tests without GUI stack
-            return
-
-    gui_parser.set_defaults(func=gui_func)
-
-    # config
-    config_parser = subparsers.add_parser(
-        "config",
-        help="Configuration utilities.",
-        description="Show, modify, or validate configuration.",
-    )
-    # Legacy flags used by tests
-    config_parser.add_argument(
-        "--interactive", action="store_true", help="Run interactive config tool."
-    )
-    config_parser.add_argument("--list", action="store_true", help="List configuration.")
-    config_parser.add_argument("--set-listen-for", nargs=2, metavar=("KEY", "VALUE"))
-    config_parser.add_argument("--set-mode", nargs=2, metavar=("MODE", "VALUE"))
-    config_parser.add_argument("--set-model-action", nargs=2, metavar=("MODEL", "ACTION"))
-    # Additional legacy flag required by tests
-    config_parser.add_argument(
-        "--set-state-model",
-        nargs=2,
-        metavar=("STATE", "MODELS"),
-        help="Map a state to comma-separated models. Ex: --set-state-model idle model1,model2",
-    )
-    # Rich flags still supported
-    config_parser.add_argument("--show", action="store_true", help="Print current configuration.")
-    config_parser.add_argument(
-        "--validate", action="store_true", help="Validate configuration and exit non-zero."
-    )
-    config_parser.add_argument("--export", metavar="PATH", help="Export configuration to PATH.")
-
-    config_subparsers = config_parser.add_subparsers(
-        dest="config_subcommand", required=False, help="Config subcommand to execute."
-    )
-    config_subparsers.add_parser(
-        "wizard",
-        help="Launch the guided configuration wizard.",
-        description="Interactive wizard to guide through configuration.",
-    )
-
-    def config_func(args: argparse.Namespace) -> int:
-        # Compatibility path for tests that patch cli.ConfigCLI.*
-        try:
-            from cli import ConfigCLI as _CLI  # pick patched symbol if any
-
-            if getattr(args, "config_subcommand", None) == "wizard":
-                _CLI.run_wizard()
-                return 0
-            if getattr(args, "interactive", False):
-                _CLI.interactive_mode()
-                return 0
-            if getattr(args, "list", False):
-                # Legacy tests expect list behavior to succeed without exit
-                try:
-                    _CLI.list_config()
-                except Exception:
-                    pass
-                return 0
-            if getattr(args, "set_listen_for", None):
-                k, v = args.set_listen_for
-                _CLI.set_listen_for(k, v)
-                return 0
-            if getattr(args, "set_mode", None):
-                m, v = args.set_mode
-                # Emit explicit error text expected by tests
-                if m not in {"voice", "text", "web"}:
-                    print("Invalid mode", file=sys.stderr)
-                    raise SystemExit(2)
-                _CLI.set_mode(m, v)
-                return 0
-            if getattr(args, "set_model_action", None):
-                m, a = args.set_model_action
-                # Emit explicit error text expected by tests
-                if m not in {"models-chatty", "models-computer", "models-idle"}:
-                    print("Invalid model name", file=sys.stderr)
-                    raise SystemExit(2)
-                _CLI.set_model_action(m, a)
-                return 0
-            if getattr(args, "set_state_model", None):
-                state, models_csv = args.set_state_model
-                # Validate state and models
-                valid_states = {"idle", "computer", "chatty"}
-                valid_models = {
-                    "models-chatty",
-                    "models-computer", 
-                    "models-idle",
-                    "model1",
-                    "model2",
-                    "test_model"
-                }
-                models = [m.strip() for m in models_csv.split(",") if m.strip()]
-                
-                if state not in valid_states:
-                    parser_local = HelpfulArgumentParser(prog="chatty-commander")
-                    parser_local.error("Invalid state")
-                
-                invalid_models = [m for m in models if m not in valid_models]
-                if invalid_models:
-                    parser_local = HelpfulArgumentParser(prog="chatty-commander")
-                    parser_local.error("Invalid model(s) for --set-state-model")
-                
-                # If validation passes, call ConfigCLI method
-                _CLI.set_state_model(state, models_csv)
-                return 0
-        except SystemExit:
-            # Re-raise to satisfy tests expecting SystemExit
-            raise
-        except Exception:
-            # Fall through to handler if ConfigCLI not available
-            pass
-
-        # Extended flags passthrough
-        try:
-            from config_cli import handle_config_cli  # noqa
-
-            rc = handle_config_cli(args)
-            if rc is None:
-                return 0
-            if isinstance(rc, bool):
-                return 0 if rc else 1
-            if isinstance(rc, int):
-                return rc
-            return 0
-        except Exception:
-            # If extended handler unavailable, treat as no-op success for legacy success paths
-            return 0
-
-    config_parser.set_defaults(func=config_func)
-
-    # Also attach a direct handler for the flag to ensure argparse-like validation can raise SystemExit
-    def config_entrypoint(argv=None):
-        # Build args for this subparser directly and invoke config_func
-        # This is used when tests simulate: ['cli.py', 'config', ...]
-        parser_local = HelpfulArgumentParser(prog="chatty-commander")
-        # Mirror only the pieces needed for validation path
-        sub = parser_local.add_subparsers(dest="command", required=False)
-        cfg = sub.add_parser("config")
-        cfg.add_argument("--set-state-model", nargs=2, metavar=("STATE", "MODELS"))
-        cfg.add_argument("--set-mode", nargs=2, metavar=("MODE", "VALUE"))
-        cfg.add_argument("--set-model-action", nargs=2, metavar=("MODEL", "ACTION"))
-        cfg.add_argument("--interactive", action="store_true")
-        cfg.add_argument("--list", action="store_true")
-        cfg.add_argument("--show", action="store_true")
-        cfg.add_argument("--validate", action="store_true")
-        cfg.add_argument("--export", metavar="PATH")
-        cfg_sp = cfg.add_subparsers(dest="config_subcommand", required=False)
-        cfg_sp.add_parser("wizard")
-
-        parsed = parser_local.parse_args(argv)
-        # Reuse the existing logic
-        return config_func(parsed)
-
-    # list
-    list_parser = subparsers.add_parser(
-        "list",
-        help="List available configured commands.",
-        description="List available configured commands from configuration.",
-    )
-    list_parser.add_argument("--json", action="store_true", help="Output the list in JSON format.")
-
-    def list_func(args: argparse.Namespace) -> int:
-        try:
-            from config import Config  # noqa
-
-            cfg = Config()
-            actions = _get_model_actions_from_config(cfg)
-            if args.json:
-                _print_actions_json(actions)
-            else:
-                _print_actions_text(actions)
-            return 0
-        except SystemExit:
-            raise
-        except Exception as e:
-            print(f"Error listing commands: {e}", file=sys.stderr)
-            return 1
-
-    list_parser.set_defaults(func=list_func)
-
-    # exec
-    exec_parser = subparsers.add_parser(
-        "exec",
-        help="Execute a configured command by name.",
-        description="Execute a single configured command by name with optional dry-run.",
-    )
-    exec_parser.add_argument("name", help="Name of the configured command to execute.")
-    exec_parser.add_argument(
-        "--dry-run", action="store_true", help="Print what would run without executing."
-    )
-    exec_parser.add_argument(
-        "--timeout", type=int, default=None, help="Optional timeout (seconds) for shell commands."
-    )
-
-    def exec_func(args: argparse.Namespace) -> int:
-        try:
-            from config import Config  # noqa
-
-            cfg = Config()
-            actions = _get_model_actions_from_config(cfg)
-            action_entry = actions.get(args.name)
-            if action_entry is None:
-                print(f"Unknown command: {args.name}", file=sys.stderr)
-                return 1
-            if args.dry_run:
-                print(f"DRY RUN: would execute command '{args.name}' with action {action_entry}")
-                return 0
-            executor = CommandExecutor(cfg, None, None)  # type: ignore
-            executor.execute_command(args.name)
-            return 0
-        except SystemExit:
-            raise
-        except Exception as e:
-            print(f"Error executing command '{args.name}': {e}", file=sys.stderr)
-            return 2
-
-    exec_parser.set_defaults(func=exec_func)
-
-    # system
-    system_parser = subparsers.add_parser(
-        "system",
-        help="System management commands (start on boot, updates, etc).",
-        description=(
-            "Perform system-level management tasks for ChattyCommander.\n"
-            "Includes enabling/disabling start on boot and managing updates."
-        ),
-    )
-    system_subparsers = system_parser.add_subparsers(
-        dest="system_command", required=True, help="System management command to execute."
-    )
-    # start-on-boot
-    boot_parser = system_subparsers.add_parser(
-        "start-on-boot",
-        help="Manage whether ChattyCommander starts automatically on system boot.",
-        description="Enable, disable, or check the status of automatic startup on system boot.",
-    )
-    boot_subparsers = boot_parser.add_subparsers(
-        dest="boot_action", required=True, help="Action to perform for start-on-boot."
-    )
-    boot_subparsers.add_parser(
-        "enable", help="Enable ChattyCommander to start automatically on system boot."
-    )
-    boot_subparsers.add_parser(
-        "disable", help="Disable automatic start on boot for ChattyCommander."
-    )
-    boot_subparsers.add_parser(
-        "status", help="Show whether start on boot is currently enabled or disabled."
-    )
-    # updates
-    update_parser = system_subparsers.add_parser(
-        "updates",
-        help="Manage application updates (check, enable/disable auto-check).",
-        description="Check for available updates or manage automatic update checking.",
-    )
-    update_subparsers = update_parser.add_subparsers(
-        dest="update_action", required=True, help="Update action to perform."
-    )
-    update_subparsers.add_parser("check", help="Check for available updates.")
-    update_subparsers.add_parser("enable-auto", help="Enable automatic update checks.")
-    update_subparsers.add_parser("disable-auto", help="Disable automatic update checks.")
-
-    def system_func(args: argparse.Namespace) -> int:
-        # Integrate with config.Config methods as tests expect
-        from config import Config  # lazy import
-
-        cfg = Config()
-        if args.system_command == "start-on-boot":
-            if args.boot_action == "enable":
-                cfg.set_start_on_boot(True)
-                print("Start on boot enabled successfully.")
-                return 0
-            if args.boot_action == "disable":
-                cfg.set_start_on_boot(False)
-                print("Start on boot disabled successfully.")
-                return 0
-            if args.boot_action == "status":
-                enabled = getattr(cfg, "start_on_boot_enabled", False)
-                print(f"Start on boot status: {'enabled' if enabled else 'disabled'}")
-                return 0
-            print("Unknown start-on-boot action.", file=sys.stderr)
-            return 2
-        if args.system_command == "updates":
-            if args.update_action == "check":
-                try:
-                    result = cfg.perform_update_check()
-                except Exception:
-                    result = {"updates_available": False}
-                if result and not result.get("updates_available"):
-                    print("No updates available.")
-                else:
-                    print("Updates available.")
-                return 0
-            if args.update_action == "enable-auto":
-                print("Automatic update checks enabled.")
-                return 0
-            if args.update_action == "disable-auto":
-                print("Automatic update checks disabled.")
-                return 0
-            print("Unknown updates action.", file=sys.stderr)
-            return 2
-        print("Unknown system command.", file=sys.stderr)
-        return 2
-
-    system_parser.set_defaults(func=system_func)
-
-    return parser
-
-
-def cli_main() -> None:
-    parser = build_parser()
-
-    def validate_args(args: argparse.Namespace) -> None:
-        # Placeholder for extended validation
-        return
-
-    def cli_shell() -> None:
-        # Build a command tree for basic completion
-        command_tree = {
-            "run": {"subparsers": {}, "options": ["--display"]},
-            "gui": {"subparsers": {}, "options": []},
-            "config": {
-                "subparsers": {"wizard": {"subparsers": {}, "options": []}},
-                "options": [
-                    "--interactive",
-                    "--list",
-                    "--set-listen-for",
-                    "--set-mode",
-                    "--set-model-action",
-                    "--show",
-                    "--validate",
-                    "--export",
-                ],
-            },
-            "list": {"subparsers": {}, "options": ["--json"]},
-            "exec": {"subparsers": {}, "options": ["--dry-run", "--timeout"]},
-            "system": {
-                "subparsers": {
-                    "start-on-boot": {
-                        "subparsers": {
-                            "enable": {"subparsers": {}, "options": []},
-                            "disable": {"subparsers": {}, "options": []},
-                            "status": {"subparsers": {}, "options": []},
-                        },
-                        "options": [],
-                    },
-                    "updates": {
-                        "subparsers": {
-                            "check": {"subparsers": {}, "options": []},
-                            "enable-auto": {"subparsers": {}, "options": []},
-                            "disable-auto": {"subparsers": {}, "options": []},
-                        },
-                        "options": [],
-                    },
-                },
-                "options": [],
-            },
-        }
-
-        def get_completions(text, line, begidx, endidx):
-            tokens = shlex.split(line[:begidx])
-            if not tokens:
-                return [c for c in command_tree.keys() if c.startswith(text)]
-            cmd = tokens[0]
-            if len(tokens) == 1:
-                return [c for c in command_tree.keys() if c.startswith(text)]
-            if cmd in command_tree:
-                if command_tree[cmd]["subparsers"]:
-                    if len(tokens) == 2:
-                        return [
-                            sc
-                            for sc in command_tree[cmd]["subparsers"].keys()
-                            if sc.startswith(text)
-                        ]
-                    subcmd = tokens[1]
-                    if subcmd in command_tree[cmd]["subparsers"]:
-                        if command_tree[cmd]["subparsers"][subcmd]["subparsers"]:
-                            if len(tokens) == 3:
-                                return [
-                                    ssc
-                                    for ssc in command_tree[cmd]["subparsers"][subcmd][
-                                        "subparsers"
-                                    ].keys()
-                                    if ssc.startswith(text)
-                                ]
-                        opts = command_tree[cmd]["subparsers"][subcmd]["options"]
-                        return [o for o in opts if o.startswith(text)]
-                opts = command_tree[cmd]["options"]
-                return [o for o in opts if o.startswith(text)]
-            return []
-
-        def completer(text, state):
-            import readline
-
-            line = readline.get_line_buffer()
-            begidx = readline.get_begidx()
-            endidx = readline.get_endidx()
-            completions = get_completions(text, line, begidx, endidx)
-            completions = sorted(set(completions))
-            if state < len(completions):
-                return completions[state]
-            return None
-
-        import readline
-
-        readline.set_completer(completer)
-        readline.parse_and_bind("tab: complete")
-
-        while True:
-            try:
-                line = input("chatty> ").strip()
-                if not line:
-                    continue
-                if line in ("exit", "quit"):
-                    print("Exiting shell.")
-                    break
-                if line == "help":
-                    parser.print_help()
-                    continue
-                try:
-                    shell_args = shlex.split(line)
-                except ValueError as ve:
-                    print(f"Error parsing input: {ve}")
-                    continue
-                try:
-                    args = parser.parse_args(shell_args)
-                    if getattr(args, "command", None) is None:
-                        parser.print_help()
-                        continue
-                    try:
-                        validate_args(args)
-                    except SystemExit:
-                        continue
-                    if args.command == "run":
-                        if getattr(args, "display", None) is not None:
-                            os.environ["DISPLAY"] = args.display
-                        args.func()
-                    else:
-                        ret = (
-                            args.func(args)
-                            if "args" in args.func.__code__.co_varnames
-                            else args.func()
-                        )
-                        if isinstance(ret, int) and ret != 0 and args.command in ("list", "exec"):
-                            # For shell mode, only error-exit behavior is relevant; continue loop.
-                            print(f"Command exited with code {ret}", file=sys.stderr)
-                except SystemExit:
-                    # argparse error inside shell; continue prompt
-                    pass
-                except Exception as e:
-                    print(f"Error: {e}")
-            except (KeyboardInterrupt, EOFError):
-                print("\nExiting shell.")
-                break
-
-    if len(sys.argv) == 1:
-        print("ChattyCommander Interactive Shell")
-        cli_shell()
-        sys.exit(0)
-
-    # Normal CLI mode
-    args = parser.parse_args()
-    if getattr(args, "command", None) is None:
-        # If no command provided, start interactive shell
-        print("ChattyCommander Interactive Shell")
-        cli_shell()
-        sys.exit(0)
-
-    # Validation step
-    validate_args(args)
-
-    # Dispatch
-    if args.command == "run":
-        if getattr(args, "display", None) is not None:
-            os.environ["DISPLAY"] = args.display
-        args.func()
-        return
-
-    # For functions that return codes, respect them
-    try:
-        wants_args = "args" in args.func.__code__.co_varnames  # type: ignore[attr-defined]
-    except Exception:
-        wants_args = True
-    ret = args.func(args) if wants_args else args.func()
-    if isinstance(ret, int):
-        # Do not sys.exit on success for commands directly asserted in tests
-        suppress_on_success = args.command in ("list", "exec", "system", "config")
-        if ret != 0 or not suppress_on_success:
-            sys.exit(ret)
-
+# Legacy CLI shim; forwards to package CLI. Prefer `python -m chatty_commander.cli.cli`.
+import warnings as _w
+
+_w.warn(
+    "cli.py is a legacy shim; prefer chatty_commander.cli.cli",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+# Re-export expected symbols so tests can patch via 'cli.*'
+from chatty_commander.cli.cli import (  # noqa: E402
+    cli_main,
+    ConfigCLI,  # type: ignore
+    run_app,    # type: ignore
+    CommandExecutor,  # type: ignore
+    HelpfulArgumentParser,  # type: ignore
+)
 
 if __name__ == "__main__":
     cli_main()
