+++ pyproject.toml
@@ -11,7 +11,6 @@ dependencies = [
     "pytest-mock>=3.12.0",
     "pytest-cov>=4.1.0",
     "requests>=2.32.4",
-    "argparse>=1.4.0", # Ensure argparse is available, though standard library
     "psutil>=7.0.0",
     "fastapi>=0.116.1",
     "uvicorn[standard]>=0.35.0",
@@ -28,7 +27,7 @@ dependencies = [
 ]
 
 [project.scripts]
-chatty-commander = "chatty_commander.cli.cli:main"
+chatty-commander = "chatty_commander.cli.cli:cli_main"
 
 [dependency-groups]
 dev = []
@@ -51,6 +50,7 @@ omit = [
     "webui/frontend/*",
     "*/__pycache__/*",
     "*/site-packages/*",
+    "config-*.py",
 ]
 
 # Removed legacy setuptools block for old flat-layout (py-modules, utils)
@@ -90,4 +90,7 @@ skip-magic-trailing-comma = false
 
 [tool.setuptools]
 package-dir = {"" = "src"}
-packages = ["chatty_commander"]
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["chatty_commander*"]
